+++ client/my-app/src/components/MusicPlayer/MusicPlayerButton.js
@@ -1,0 +1,42 @@
+import {changeOnHover, changeOnMouseDown, changeOnMouseOut, changeOnMouseUp, PageContainer} from "../../utils/utils";
+import pausebutton from "../../images/pause.png";
+import playbutton from "../../images/play.png";
+import thankyou from "../../sounds/thankyou.mp3";
+import React, {useContext} from "react";
+import {AudioContext} from "../AudioContext/AudioContext";
+
+export const MusicPlayerButton = () => {
+    const { play, pause, isPlaying, currentTrack } = useContext(AudioContext);
+    const handlePlay = (track) => {
+        play(track);
+        console.log(currentTrack);
+    };
+    const handlePause = (track) => {
+        pause(track);
+        console.log(currentTrack);
+    };
+    return (
+        <div>
+            <img
+                id="playButton"
+                alt="playbutton"
+                src={isPlaying ? pausebutton : playbutton}
+                style={{height: "80px", width: "80px", right: "0px", bottom: "0px", position: "fixed"}}
+                onClick={() => {
+                    if (!isPlaying) {
+                        handlePlay({ url: thankyou});
+                        document.getElementById("playbutton").src = pausebutton;
+                    }
+                    else if (isPlaying) {
+                        handlePause({ url: thankyou});
+                        document.getElementById("playbutton").src = playbutton;
+                    }
+                }}
+                onMouseOver={() => changeOnHover("playButton", isPlaying)}
+                onMouseOut={() => changeOnMouseOut("playButton", isPlaying)}
+                onMouseUp={() => changeOnMouseUp("playButton", isPlaying)}
+                onMouseDown={() => changeOnMouseDown("playButton", isPlaying)}
+            ></img>
+        </div>
+    );
+};
