+++ client/my-app/src/pages/Contact/Contact.js
@@ -3,12 +3,14 @@ import styled from 'styled-components';
 import GlobalHeader from '../../components/GlobalHeader/GlobalHeader';
 import GlobalFooter from '../../components/GlobalFooter/GlobalFooter';
 import { PageContainer } from '../../utils/utils';
+import {MusicPlayerButton} from "../../components/MusicPlayer/MusicPlayerButton";
 
 const Contact = () => {
   return (
     <PageContainer>
       <GlobalHeader />
-      <Content>
+        <MusicPlayerButton  />
+        <Content>
       </Content>
       <GlobalFooter />
     </PageContainer>
